[Enzo] Partially fixed the node creation issue. There is still an issue with the entityManager whe I try to create a NodeHU, it works with all other (not tested with NodeCU).

+++ ProjetMinuteMed/src/main/java/fr/devsquad/minutemed/arborescence/rest/ArborescenceService.java
@@ -14,14 +14,16 @@ import io.swagger.annotations.ApiResponses;
 import java.util.*;
 import javax.ejb.*;
 import javax.inject.*;
+import javax.validation.*;
 import javax.validation.constraints.NotNull;
 import javax.ws.rs.*;
+import javax.ws.rs.Path;
 import javax.ws.rs.core.*;
 
 @Path("/nodes")
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
-@JWTNeeded(groups = {StaffEnum.DATA_MANAGER, StaffEnum.DOCTOR})
+//@JWTNeeded(groups = {StaffEnum.DATA_MANAGER})
 @Api("Arborescence REST Endpoint")
 public class ArborescenceService {
     
@@ -46,16 +48,16 @@ public class ArborescenceService {
         @ApiResponse(code = 400, message = "Invalid input")}
     )
     @Path("/APHP")
-    public Response createAPHP(@NotNull NodeAPHP aphp) {
-        try{
-            repository.findNode(new Long(1), NodeAPHP.class);
-        }catch(EJBException ex){
-            Long id = repository.saveNode(aphp);
+    public Response createAPHP() {
+        NodeAPHP aphp = repository.findNode(51L, NodeAPHP.class);
+        if(aphp == null){
+            Long id = repository.saveNode(new NodeAPHP());
             return Response.ok("{\"idAPHPCreated\":"+ id +"}").build();
         }
-        return Response.status(Response.Status.BAD_REQUEST)
+        return Response.status(Response.Status.CONFLICT)
                 .entity("APHP already exist !").build();
     }
+    
     
     @POST
     @ApiOperation(value = "Create an Hospital")
@@ -63,24 +65,27 @@ public class ArborescenceService {
         @ApiResponse(code = 201, message = "The Hospital is created."),
         @ApiResponse(code = 400, message = "Invalid input")}
     )
-    @Path("/APHP/hospitals")
-    public Response createHospital(@NotNull NodeHospital hospital) {
-        NodeAPHP aphp = hospital.getFather();
+    @Path("/APHP/{idAPHP}/hospitals")
+    public Response createHospital(@PathParam("idAPHP") Long idAPHP) {
+        NodeAPHP aphp = repository.findNode(idAPHP, NodeAPHP.class);
         if(aphp == null){
             return Response.status(Response.Status.BAD_REQUEST)
-                    .entity("The APHP node is required in the Hospital node !")
+                    .entity("This APHP node does not exist !")
                     .build();
         }
+        System.out.println(aphp.getIdNode());
+        NodeHospital hospital = new NodeHospital(aphp);
         Long id = repository.saveNode(hospital);
+        System.out.println(id);
         if(!aphp.addHospital(hospital)){
             repository.removeNode(hospital);
             return Response.status(Response.Status.BAD_REQUEST)
                     .entity("This Hospital already exist in the APHP !")
                     .build();           
         }
-        repository.refreshNode(aphp);
-        return Response.ok("{\"idHospitalCreated\":"+ aphp.getIdNode() +"}").build();
+        return Response.ok("{\"idHospitalCreated\":"+ id +"}").build();
     }
+    
     
     @POST
     @ApiOperation(value = "Create a Pole")
@@ -89,19 +94,20 @@ public class ArborescenceService {
         @ApiResponse(code = 400, message = "Invalid input."),
         @ApiResponse(code = 404, message = "Unknow node.")}
     )
-    @Path("/APHP/hospitals/{idHospital}/poles")
-    public Response createPole(@PathParam("idHospital") Long idHospital, @NotNull NodePole pole) {
-        NodeHospital hospital = pole.getFather();
+    @Path("/APHP/{idAPHP}/hospitals/{idHospital}/poles")
+    public Response createPole(@PathParam("idAPHP") Long idAPHP, @PathParam("idHospital") Long idHospital) {
+        NodeHospital hospital = repository.findNode(idHospital, NodeHospital.class);
         if(hospital == null){
             return Response.status(Response.Status.BAD_REQUEST)
                     .entity("The Hospital node is required in the Pole node !")
                     .build();
         }
-        if(hospital.getIdNode() != idHospital){
+        if(hospital.getFather().getIdNode() != idAPHP){
             return Response.status(Response.Status.BAD_REQUEST)
-                    .entity("The idHospital param is not equal to the father id !")
-                    .build();    
+                    .entity("The idAPHP param is not equal to the father id !")
+                    .build();            
         }
+        NodePole pole = new NodePole(hospital);
         Long id = repository.saveNode(pole);
         if(!hospital.addPole(pole)){
             repository.removeNode(pole);
@@ -109,9 +115,10 @@ public class ArborescenceService {
                     .entity("This Pole already exist in the Hospital !")
                     .build();           
         }
-        repository.refreshNode(hospital);
+        //repository.refreshNode(hospital);
         return Response.ok("{\"idPoleCreated\":"+ id +"}").build();
     }
+    
     
     @POST
     @ApiOperation(value = "Create a Service")
@@ -120,9 +127,9 @@ public class ArborescenceService {
         @ApiResponse(code = 400, message = "Invalid input."),
         @ApiResponse(code = 404, message = "Unknow node.")}
     )
-    @Path("/APHP/hospitals/{idHospital}/poles/{idPole}/services")
-    public Response createService(@PathParam("idHospital") Long idHospital, @PathParam("idPole") Long idPole, @NotNull NodeService service) {
-        NodePole pole = service.getFather();
+    @Path("/APHP/{idAPHP}/hospitals/{idHospital}/poles/{idPole}/services")
+    public Response createService(@PathParam("idAPHP") Long idAPHP, @PathParam("idHospital") Long idHospital, @PathParam("idPole") Long idPole) {
+        NodePole pole = repository.findNode(idPole, NodePole.class);
         if(pole == null){
             return Response.status(Response.Status.BAD_REQUEST)
                     .entity("The Pole node is required in the Service node !")
@@ -138,6 +145,12 @@ public class ArborescenceService {
                     .entity("The idHospital param is not equal to the father id !")
                     .build(); 
         }
+        if(pole.getFather().getFather().getIdNode() != idAPHP){
+            return Response.status(Response.Status.BAD_REQUEST)
+                    .entity("The idAPHP param is not equal to the father id !")
+                    .build(); 
+        }
+        NodeService service = new NodeService(pole);
         Long id = repository.saveNode(service);
         if(!pole.addService(service)){
             repository.removeNode(service);
@@ -145,9 +158,10 @@ public class ArborescenceService {
                     .entity("This Service already exist in the Pole !")
                     .build();           
         }
-        repository.refreshNode(pole);
+        //repository.refreshNode(pole);
         return Response.ok("{\"idServiceCreated\":"+ id +"}").build();
     }
+    
     
     @POST
     @ApiOperation(value = "Create a Hospital Unit")
@@ -156,10 +170,10 @@ public class ArborescenceService {
         @ApiResponse(code = 400, message = "Invalid input"),
         @ApiResponse(code = 404, message = "Unknow node.")}
     )
-    @Path("/APHP/hospitals/{idHospital}/poles/{idPole}/services/{idService}/hUnits")
-    public Response createHospitalUnit(@PathParam("idHospital") Long idHospital, @PathParam("idPole") Long idPole,
-            @PathParam("idService") Long idService, @NotNull NodeHU hu) {
-        NodeService service = hu.getFather();
+    @Path("/APHP/{idAPHP}/hospitals/{idHospital}/poles/{idPole}/services/{idService}/hUnits")
+    public Response createHospitalUnit(@PathParam("idAPHP") Long idAPHP, @PathParam("idHospital") Long idHospital,
+            @PathParam("idPole") Long idPole, @PathParam("idService") Long idService) {
+        NodeService service = repository.findNode(idService, NodeService.class);
         if(service == null){
             return Response.status(Response.Status.BAD_REQUEST)
                     .entity("The Service node is required in the Hospital Unit node !")
@@ -180,16 +194,28 @@ public class ArborescenceService {
                     .entity("The idHospital param is not equal to the father id !")
                     .build(); 
         }
-        Long id = repository.saveNode(hu);
+        if(service.getFather().getFather().getFather().getIdNode() != idAPHP){
+            return Response.status(Response.Status.BAD_REQUEST)
+                    .entity("The idAPHP param is not equal to the father id !")
+                    .build(); 
+        }
+        NodeHU hu = new NodeHU(service);
+        try{
+            Long id = repository.saveNode(hu);
+        }catch(EJBException cve){
+            return Response.status(Response.Status.SERVICE_UNAVAILABLE).entity("J'ai pas compris..").build();
+        }
+        /*
         if(!service.addHospitalUnit(hu)){
             repository.removeNode(hu);
             return Response.status(Response.Status.BAD_REQUEST)
                     .entity("This Hospital Unit already exist in the Service !")
                     .build();           
-        }
-        repository.refreshNode(service);
-        return Response.ok("{\"idHUCreated\":"+ id +"}").build();
+        }*/
+        //repository.refreshNode(service);
+        return Response.status(Response.Status.CREATED).entity("{\"idHUCreated\":"+ 101 +"}").build();
     }
+    
     
     @POST
     @ApiOperation(value = "Create a Care Unit")
@@ -198,8 +224,8 @@ public class ArborescenceService {
         @ApiResponse(code = 400, message = "Invalid input"),
         @ApiResponse(code = 404, message = "Unknow node.")}
     )
-    @Path("/APHP/hospitals/{idHospital}/poles/{idPole}/services/{idService}/hUnits/{idHU}/cUnits")
-    public Response createCareUnit(@PathParam("idHospital") Long idHospital, @PathParam("idPole") Long idPole,
+    @Path("/APHP/{idAPHP}/hospitals/{idHospital}/poles/{idPole}/services/{idService}/hUnits/{idHU}/cUnits")
+    public Response createCareUnit(@PathParam("idAPHP") Long idAPHP, @PathParam("idHospital") Long idHospital, @PathParam("idPole") Long idPole,
             @PathParam("idService") Long idService, @PathParam("idHU") Long idHU, @NotNull NodeCU cu) {
         NodeHU hospitalUnit = cu.getFather();
         if(hospitalUnit == null){
@@ -227,6 +253,11 @@ public class ArborescenceService {
                     .entity("The idHospital param is not equal to the father id !")
                     .build(); 
         }
+        if(hospitalUnit.getFather().getFather().getFather().getFather().getIdNode() != idAPHP){
+            return Response.status(Response.Status.BAD_REQUEST)
+                    .entity("The idAPHP param is not equal to the father id !")
+                    .build(); 
+        }
         Long id = repository.saveNode(cu);
         if(!hospitalUnit.addCareUnit(cu)){
             repository.removeNode(cu);
@@ -234,7 +265,7 @@ public class ArborescenceService {
                     .entity("This Care Unit already exist in the Hospital Unit !")
                     .build();           
         }
-        repository.refreshNode(hospitalUnit);
+        //repository.refreshNode(hospitalUnit);
         return Response.ok("{\"idCUCreated\":"+ id +"}").build();
     }
     
@@ -253,8 +284,11 @@ public class ArborescenceService {
     )
     public Response getAPHP() {
         
-        NodeAPHP aphp = repository.findNode(1L, NodeAPHP.class);
-        return Response.ok(aphp).build();
+        List<NodeAPHP> aphps = repository.findNodes(NodeAPHP.class);
+        if(aphps.isEmpty()){
+            return Response.status(Response.Status.NOT_FOUND).build();
+        }
+        return Response.ok(aphps.get(0)).build();
     }
     
     @GET
@@ -400,12 +434,8 @@ public class ArborescenceService {
         @ApiResponse(code = 400, message = "Invalid input")}
     )
     @JWTNeeded(groups = {StaffEnum.DOCTOR})
-    public Response getAccessibleCareUnits(@Context HttpHeaders headers) {
-        List<String> authorizations = headers.getRequestHeader(HttpHeaders.AUTHORIZATION);
-        authorizations.stream().forEach(System.out::println);
-        System.out.println("getAccessibleCareUnits");
-        String authorization = authorizations.get(0);
-        String token = authorization.substring("Bearer".length()).trim();
+    public Response getAccessibleCareUnits(@HeaderParam(HttpHeaders.AUTHORIZATION) String authorizationHeader) {
+        String token = authorizationHeader.substring("Bearer".length()).trim();
         MedicalStaff user = staffRepository.findMedicalStaff(tokenUtils.decryptIdFromToken(token));
         Set<NodeCU> careUnits = repository.findNode(user.getNode().getIdNode(), Node.class).getAccessibleNode();
         return Response.ok(careUnits).build();
